requirements.txt
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
scipy>=1.9.0
pyyaml>=6.0
requests>=2.28.0
joblib>=1.2.0  # ← ADICIONAR esta linha